+++ posture.h
@@ -12,28 +12,28 @@ void posture_exit(void);
 void posture_reset(void);
 
 //获取角速度计算的转角(相对于自身坐标系)
-float posture_getAGX(void);
-float posture_getAGY(void);
-float posture_getAGZ(void);
+double posture_getAGX(void);
+double posture_getAGY(void);
+double posture_getAGZ(void);
 
 //获取重力加速度计算的转角(相对于空间坐标系)
-float posture_getACX(void);
-float posture_getACY(void);
-float posture_getACZ(void);
+double posture_getACX(void);
+double posture_getACY(void);
+double posture_getACZ(void);
 
 //最终输出转角
-float posture_getX(void);
-float posture_getY(void);
-float posture_getZ(void);
+double posture_getX(void);
+double posture_getY(void);
+double posture_getZ(void);
 
 //获取加速度计数据
-short posture_getAccelX(void);
-short posture_getAccelY(void);
-short posture_getAccelZ(void);
+int posture_getAccelX(void);
+int posture_getAccelY(void);
+int posture_getAccelZ(void);
 
 //获取陀螺仪数据
-short posture_getGyroX(void);
-short posture_getGyroY(void);
-short posture_getGyroZ(void);
+int posture_getGyroX(void);
+int posture_getGyroY(void);
+int posture_getGyroZ(void);
 
 #endif
